Fixed VUSkip a bit - it is no longer unstable and should also give better/more consistent skipping then before.
MTGS mode fix: Retooled the MTGS's GIFtag handler so that it's based on Gsdx's GIFtag handler (may fix MTGS problems in DigitalDevilSaga).

Updated VM build to give an error when trying to load VTLB states.

Various tweaks and code cleanups.

git-svn-id: http://pcsx2-playground.googlecode.com/svn/trunk@467 a6443dda-0b58-4228-96e9-037be469359c

+++ pcsx2/FiFo.c
@@ -131,7 +131,7 @@ void WriteFIFO(u32 mem, const u64 *value) {
 
 			data[0] = value[0];
 			data[1] = value[1];
-			GSgifTransferDummy(2, (u32*)data, 1);
+			GSgifTransferDummy(2, (u8*)data, 1);
 			GSRINGBUF_DONECOPY((u8*)data, 16);
 		}
 		else {
